moving all output interface to main.rb
